osiris/requirements.txt
Django==2.2.2
django-rest-framework
psycopg2
gunicorn
coverage
schema
psutil
matplotlib
websockets
Pillow
tensorflow
scikit-learn
numpy